prevent rank() integer overflow
git-svn-id: https://svn.r-project.org/R/trunk@85617 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/sort.c
@@ -1570,7 +1570,7 @@ attribute_hidden SEXP do_rank(SEXP call, SEXP op, SEXP args, SEXP rho)
 		switch(ties_kind) {
 		case AVERAGE:
 		    for (k = i; k <= j; k++)
-			rk[in[k]] = (i + j + 2) / 2.;
+			rk[in[k]] = ((double)i + j + 2) / 2.;
 		    break;
 		case MAX:
 		    for (k = i; k <= j; k++) ik[in[k]] = j+1;
